+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `previousElapsedDurationSeconds` to `NewDraftSubmission`
+
 ## [23.6.0] - 2025-12-04
 
 ### Dependencies